Update versions for tarantool modules
This patch updates the version for each tarantool module to the latest
tag of the module where it's possible.

Also, a build issue was fixed for Centos 7. It was impossible to
download the pgdg-redhat-repo-latest.noarch.rpm package by the link
from the docker file. So the link was updated.

Clases #208
Closes #215