Update the version for mqtt module
This patch updates the version for mqtt module to the latest tag and
removes the steps with installing the mosquitto library for alpine:3.9.
With the new release of mqtt module, libmosquitto builds statically
when the module is installed from luarocks. That fix eliminates the
dependencies that are not necessary anymore.

Closes tarantool/mqtt#41