Replace variable LUAJIT_ENABLE_GC64 to GC64 for building images
  * In alpine images used to LUAJIT_ENABLE_GC64 for enable GC64 now just GC64.
  * In centos and ubuntu images didn't GC64 aplied, fixed.

Default value GC64 set in .gitlab.mk to OFF and can be changed to ON, for centos
and ubuntu tarantool install from script wich use value GC64 true or false
that's why in docker files for ubuntu centos present if clause for changing OFF
to false or ON to true. Default value GC64 set in .gitlab.mk to OFF and can be
changed, for centos and ubuntu.